+++ packages/client-auth/RELEASES.md
@@ -109,3 +109,13 @@ Based on:
 - [typescript v1.0.0-alpha.10] packages/client-auth
 ### Releases
 - [NPM v1.0.0-alpha.10] https://www.npmjs.com/package/@formance/sdk-auth/v/1.0.0-alpha.10 - packages/client-auth
+
+## 2024-07-23 00:09:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.345.1 (2.378.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.11] packages/client-auth
+### Releases
+- [NPM v1.0.0-alpha.11] https://www.npmjs.com/package/@formance/sdk-auth/v/1.0.0-alpha.11 - packages/client-auth